pyproject.toml
[tool.poetry]
name = "payment-gateway-clean"
version = "0.1.0"
description = ""
authors = ["renaldyhidayatt <renaldyhidayatt@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
alembic = "^1.14.0"
asyncpg = "^0.30.0"
fastapi = "^0.115.5"
bcrypt = "^4.2.1"
pyjwt = "^2.10.0"
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
sqlalchemy-utils = "^0.41.2"
uvicorn = "^0.32.1"
structlog = "^24.4.0"
pydantic = {extras = ["email"], version = "^2.10.2"}
pydantic-settings = "^2.6.1"
psycopg2-binary = "^2.9.10"
dependency-injector = "^4.43.0"
python-jose = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
